requirements.txt
streamlit>=1.30
python-dotenv>=1.0
google-generativeai>=0.7